analyzer/requirements.txt
psycopg2==2.8.5
sqlalchemy==1.3.17
nltk==3.5
goslate==1.5.1
python-dotenv==0.13.0
vaderSentiment>=3.3.1